Migrate drop down into separate property editors for single and multiple
