+++ src/Umbraco.Core/PropertyEditors/SingleDropDownConfiguration.cs
@@ -1,0 +1,8 @@
+namespace Umbraco.Cms.Core.PropertyEditors;
+
+/// <summary>
+///     Represents the configuration for the dropdown property editor holding a single value.
+/// </summary>
+public class SingleDropDownConfiguration : ValueListConfiguration
+{
+}
