+++ src/Umbraco.Infrastructure/PropertyEditors/SingleDropDownPropertyEditor.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Umbraco.
+// Copyright (c) Umbraco.
 // See LICENSE for more details.
 
 using Umbraco.Cms.Core.IO;
@@ -7,23 +7,27 @@ using Umbraco.Cms.Core.Serialization;
 namespace Umbraco.Cms.Core.PropertyEditors;
 
 /// <summary>
-/// Represents a drop-down property editor holding any number of the configured values.
+/// Represents a drop-down property editor holding a single one of the configured values.
 /// </summary>
+/// <remarks>
+/// Stores the same value as <see cref="DropDownFlexiblePropertyEditor" /> and is edited the same way; the two
+/// differ in the shape of the value they yield, and in how many values may be selected.
+/// </remarks>
 [DataEditor(
-    Constants.PropertyEditors.Aliases.MultipleDropDown,
+    Constants.PropertyEditors.Aliases.SingleDropDown,
     ValueEditorIsReusable = true)]
-public class DropDownFlexiblePropertyEditor : DropDownPropertyEditorBase
+public class SingleDropDownPropertyEditor : DropDownPropertyEditorBase
 {
     private readonly IIOHelper _ioHelper;
     private readonly IConfigurationEditorJsonSerializer _configurationEditorJsonSerializer;
 
     /// <summary>
-    /// Initializes a new instance of the <see cref="DropDownFlexiblePropertyEditor"/> class.
+    /// Initializes a new instance of the <see cref="SingleDropDownPropertyEditor"/> class.
     /// </summary>
     /// <param name="dataValueEditorFactory">Factory used to create data value editors for property values.</param>
     /// <param name="ioHelper">Helper for IO operations, such as file and path handling.</param>
     /// <param name="configurationEditorJsonSerializer">Serializer for handling JSON configuration of the editor.</param>
-    public DropDownFlexiblePropertyEditor(IDataValueEditorFactory dataValueEditorFactory, IIOHelper ioHelper, IConfigurationEditorJsonSerializer configurationEditorJsonSerializer)
+    public SingleDropDownPropertyEditor(IDataValueEditorFactory dataValueEditorFactory, IIOHelper ioHelper, IConfigurationEditorJsonSerializer configurationEditorJsonSerializer)
         : base(dataValueEditorFactory)
     {
         _ioHelper = ioHelper;
@@ -32,5 +36,5 @@ public class DropDownFlexiblePropertyEditor : DropDownPropertyEditorBase
 
     /// <inheritdoc />
     protected override IConfigurationEditor CreateConfigurationEditor() =>
-        new DropDownFlexibleConfigurationEditor(_ioHelper, _configurationEditorJsonSerializer);
+        new SingleDropDownConfigurationEditor(_ioHelper, _configurationEditorJsonSerializer);
 }
