+++ tests/Umbraco.Tests.Integration/Umbraco.Search.Provider.Examine/Tests/ContentTests/SearchService/InvariantFacetsTests.cs
@@ -1,4 +1,4 @@
-using System.Globalization;
+﻿using System.Globalization;
 using Microsoft.Extensions.DependencyInjection;
 using NUnit.Framework;
 using Umbraco.Cms.Core;
@@ -360,7 +360,7 @@ public class InvariantFacetsTests : SearcherTestBase
             .AddPropertyType()
             .WithAlias("dropDown")
             .WithDataTypeId(Constants.DataTypes.DropDownSingle)
-            .WithPropertyEditorAlias(Constants.PropertyEditors.Aliases.DropDownListFlexible)
+            .WithPropertyEditorAlias(Constants.PropertyEditors.Aliases.MultipleDropDown)
             .Done()
             .Build();
         await ContentTypeService.CreateAsync(ContentType, Constants.Security.SuperUserKey);
